+++ CHANGELOG.md
@@ -38,6 +38,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Downloadable ZIP files of badge images for each system from Mission Control
 - Ability to attach, remove, or update a certificate (uploaded PDF) to a trigger
 - News stand post updates are shown as updated and sorted by update date
+- Uploading a list of vendor codes to invalidate them and send participants new ones
 
 ### Fixed